+++ rust/flight/src/bin/main.rs
@@ -1,0 +1,37 @@
+// Licensed to the Apache Software Foundation (ASF) under one
+// or more contributor license agreements.  See the NOTICE file
+// distributed with this work for additional information
+// regarding copyright ownership.  The ASF licenses this file
+// to you under the Apache License, Version 2.0 (the
+// "License"); you may not use this file except in compliance
+// with the License.  You may obtain a copy of the License at
+//
+//   http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing,
+// software distributed under the License is distributed on an
+// "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY
+// KIND, either express or implied.  See the License for the
+// specific language governing permissions and limitations
+// under the License.
+
+pub mod arrow_flight {
+    tonic::include_proto!("arrow.flight.protocol"); // The string specified here must match the proto package name
+}
+use arrow_flight::server::{FlightService, FlightServiceServer};
+use tonic::transport::Server;
+
+use flight::FlightServiceImpl;
+
+#[tokio::main]
+async fn main() -> Result<(), Box<dyn std::error::Error>> {
+    let addr = "[::1]:50051".parse()?;
+    let service = FlightServiceImpl::new();
+
+    Server::builder()
+        .add_service(FlightServiceServer::new(service))
+        .serve(addr)
+        .await?;
+
+    Ok(())
+}
