+++ rust/flight/src/lib.rs
@@ -15,4 +15,94 @@
 // specific language governing permissions and limitations
 // under the License.
 
-pub mod server;
+use tokio::sync::mpsc;
+use tonic::codegen::*;
+use tonic::{transport::Server, Request, Response, Status, Streaming};
+
+pub mod arrow_flight {
+    tonic::include_proto!("arrow.flight.protocol"); // The string specified here must match the proto package name
+}
+use arrow_flight::{
+    server, server::FlightService, server::FlightServiceServer, Action, ActionType,
+    Criteria, Empty, FlightData, FlightDescriptor, FlightInfo, HandshakeRequest,
+    HandshakeResponse, PutResult, SchemaResult, Ticket,
+};
+
+pub struct FlightServiceImpl {}
+
+impl FlightServiceImpl {
+    pub fn new() -> Self {
+        Self {}
+    }
+}
+
+#[tonic::async_trait]
+impl FlightService for FlightServiceImpl {
+    type HandshakeStream = mpsc::Receiver<Result<HandshakeResponse, Status>>;
+
+    async fn handshake(
+        &self,
+        request: Request<Streaming<HandshakeRequest>>,
+    ) -> Result<Response<Self::HandshakeStream>, Status> {
+        unimplemented!()
+    }
+
+    type ListFlightsStream = mpsc::Receiver<Result<FlightInfo, Status>>;
+
+    async fn list_flights(
+        &self,
+        request: Request<Criteria>,
+    ) -> Result<Response<Self::ListFlightsStream>, Status> {
+        unimplemented!()
+    }
+
+    async fn get_flight_info(
+        &self,
+        request: Request<FlightDescriptor>,
+    ) -> Result<Response<FlightInfo>, Status> {
+        unimplemented!()
+    }
+
+    async fn get_schema(
+        &self,
+        request: Request<FlightDescriptor>,
+    ) -> Result<Response<SchemaResult>, Status> {
+        unimplemented!()
+    }
+
+    type DoGetStream = mpsc::Receiver<Result<FlightData, Status>>;
+
+    async fn do_get(
+        &self,
+        request: Request<Ticket>,
+    ) -> Result<Response<Self::DoGetStream>, Status> {
+        unimplemented!()
+    }
+
+    type DoPutStream = mpsc::Receiver<Result<PutResult, Status>>;
+
+    async fn do_put(
+        &self,
+        request: Request<Streaming<FlightData>>,
+    ) -> Result<Response<Self::DoPutStream>, Status> {
+        unimplemented!()
+    }
+
+    type DoActionStream = mpsc::Receiver<Result<arrow_flight::Result, Status>>;
+
+    async fn do_action(
+        &self,
+        request: Request<Action>,
+    ) -> Result<Response<Self::DoActionStream>, Status> {
+        unimplemented!()
+    }
+
+    type ListActionsStream = mpsc::Receiver<Result<ActionType, Status>>;
+
+    async fn list_actions(
+        &self,
+        request: Request<Empty>,
+    ) -> Result<Response<Self::ListActionsStream>, Status> {
+        unimplemented!()
+    }
+}
